Refactor listings into listing/types
