+++ backend/core/index.ts
@@ -2,3 +2,9 @@
 export * from "./src/user/entities/user.entity"
 export * from "./src/listings/entities/listing.entity"
 export * from "./src/units/entities/unit.entity"
+export { ListingEventType } from "./src/listings/types/listing-event-type-enum"
+export { ListingStatus } from "./src/listings/types/listing-status-enum"
+export { Asset } from "./src/listings/types/asset"
+export { ListingEvent } from "./src/listings/types/listing-event"
+export { ApplicationMethodType } from "./src/listings/types/application-method-type-enum.ts"
+export { ApplicationMethod } from "./src/listings/types/application-method"
