+++ backend/core/listings-importer.ts
@@ -2,7 +2,7 @@ import * as client from "./client"
 import axios from "axios"
 import { serviceOptions } from "./client"
 import fs from "fs"
-import { ListingStatus } from "./src/listings/entities/listing.entity"
+import { ListingStatus } from "./src/listings/types/listing-status-enum"
 
 // NOTE: unit's monthlyRent, floor and monthlyIncomeMin have been changed to type string
 // NOTE: in the DB and unit_transformations lib has been adjusted (verify correctness)
