+++ backend/core/src/listings/types/application-method-type-enum.ts.ts
@@ -1,0 +1,8 @@
+export enum ApplicationMethodType {
+  Internal = "Internal",
+  FileDownload = "FileDownload",
+  ExternalLink = "ExternalLink",
+  PaperPickup = "PaperPickup",
+  POBox = "POBox",
+  LeasingAgent = "LeasingAgent",
+}
