+++ backend/core/src/listings/types/application-method.ts
@@ -1,0 +1,28 @@
+import { Expose } from "class-transformer"
+import { IsBoolean, IsEnum, IsOptional, IsString } from "class-validator"
+import { ValidationsGroupsEnum } from "../../shared/validations-groups.enum"
+import { ApplicationMethodType } from "./application-method-type-enum.ts"
+import { ApiProperty } from "@nestjs/swagger"
+
+export class ApplicationMethod {
+  @Expose()
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  @IsEnum(ApplicationMethodType, { groups: [ValidationsGroupsEnum.default] })
+  @ApiProperty({ enum: ApplicationMethodType, enumName: "ApplicationMethodType" })
+  type: ApplicationMethodType
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  label: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  externalReference: string | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsBoolean({ groups: [ValidationsGroupsEnum.default] })
+  acceptsPostmarkedApplications: boolean | null
+}
