+++ backend/core/src/listings/types/asset.ts
@@ -1,0 +1,13 @@
+import { Expose } from "class-transformer"
+import { IsString } from "class-validator"
+import { ValidationsGroupsEnum } from "../../shared/validations-groups.enum"
+
+export class Asset {
+  @Expose()
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  label: string
+
+  @Expose()
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  fileId: string
+}
