+++ backend/core/src/listings/types/listing-event-type-enum.ts
@@ -1,0 +1,5 @@
+export enum ListingEventType {
+  openHouse = "openHouse",
+  publicLottery = "publicLottery",
+  lotteryResults = "lotteryResults",
+}
