+++ backend/core/src/listings/types/listing-status-enum.ts
@@ -1,0 +1,4 @@
+export enum ListingStatus {
+  active = "active",
+  pending = "pending",
+}
